feat: update type uint64 to int64

+++ service/user/config/debug.go
@@ -2,5 +2,5 @@ package config
 
 type Debug struct {
 	EnableMockLogin bool
-	MockUserId      int
+	MockUserId      int64
 }
